% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuCor.R
\name{AuCor}
\alias{AuCor}
\title{Function to obtain temporal autocorrelation}
\usage{
AuCor(x, l)
}
\arguments{
\item{x}{A vector of values}

\item{l}{The lag used to investigate the Autocorrelation}
}
\description{
Function to obtain temporal autocorrelation
}
