% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bi.Hist.R
\name{Bi.Hist}
\alias{Bi.Hist}
\title{Function to get a univariate mozaic plot}
\usage{
Bi.Hist(Data, Var, y)
}
\arguments{
\item{Data}{A data frame with the variables}

\item{Var}{a string with the name of the variable}

\item{y}{a string of the name of response variable}
}
\description{
Function to get a univariate mozaic plot
}
