% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiMozaic.R
\name{BiMozaic}
\alias{BiMozaic}
\title{Function to get a univariate mozaic plot}
\usage{
BiMozaic(Data, x, y, palette = "Set1")
}
\arguments{
\item{Data}{A data frame with the variables}

\item{x}{a string with the name of the variable}

\item{y}{a string of the name of response variable}

\item{palette}{color palette for the plot}
}
\description{
Function to get a univariate mozaic plot
}
