% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiV.Tbl.R
\name{BiV.Tbl}
\alias{BiV.Tbl}
\title{Function to get a Bivariate table}
\usage{
BiV.Tbl(Data, x, y)
}
\arguments{
\item{Data}{A data frame with the variables}

\item{x}{a string with the name of the variable}

\item{y}{the name of the response variable}
}
\description{
Function to get a Bivariate table
}
