% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivar.tbl.R
\name{Bivar.tbl}
\alias{Bivar.tbl}
\title{Function to get a univariate mozaic plot}
\usage{
Bivar.tbl(Data, x, y, as_kable = F)
}
\arguments{
\item{Data}{A data frame with the variables}

\item{x}{a string with the name of the variable}

\item{y}{a string of the name of response variable}

\item{as_kable}{Is the outcome table a kable format?}
}
\description{
Function to get a univariate mozaic plot
}
