% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisgregateHx.R
\name{DisgregateHx}
\alias{DisgregateHx}
\title{Function to obtain temporal autocorrelation}
\usage{
DisgregateHx(shp, Hx, field, fun = "max", res = 1000)
}
\arguments{
\item{shp}{The shapefile with the variable to convert to a grid}

\item{Hx}{The hexagonal grid shapefile, must be same crs than the shapefile}

\item{field}{The name of the field that want to obtain}

\item{fun}{A function to apply to summarise the values}

\item{res}{the resolution of the raster used to summarise the poligons}
}
\description{
Function to obtain temporal autocorrelation
}
