% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers.R
\name{fishers}
\alias{fishers}
\title{Function to obtain a Fisher's exact table}
\usage{
fishers(
  data,
  x,
  y,
  offset = 0.01,
  outcome = "figure",
  pal = c("orchid3", "seagreen3")
)
}
\arguments{
\item{data}{The name of the data frame}

\item{x}{Name of variable}

\item{y}{Name of respose variable}

\item{offset}{spacing between the mosaic in the figure}

\item{outcome}{Type of outcome to return, options are 'figure' and 'table'}

\item{pal}{color palette for the figure. Must provide two or more colors, function automatically interpolates between colors if number of colors provided less than levels of y variable}
}
\description{
Function to obtain a Fisher's exact table
}
