% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishersTbl.R
\name{FishersTbl}
\alias{FishersTbl}
\title{Function to obtain a Fisher's exact table}
\usage{
FishersTbl(Data, x, y)
}
\arguments{
\item{Data}{The name of the data frame}

\item{x}{Name of variable}

\item{y}{Name of respose variable}
}
\description{
Function to obtain a Fisher's exact table
}
