% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_coords.R
\name{grid_coords}
\alias{grid_coords}
\title{Function to generate coordinates in a 2D grid}
\usage{
grid_coords(data, w, order = "ordered")
}
\arguments{
\item{data}{a dataset where each row will have a coordinate in a 2d plane assigned}

\item{w}{Name of the variable used to generate the coordinates}

\item{order}{order of the arrangement}
}
\description{
Function to generate coordinates in a 2D grid
}
