% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hCol.R
\name{hCol}
\alias{hCol}
\title{Function for color based on Hue, Brightness and Sat}
\usage{
hCol(h, b, s)
}
\arguments{
\item{h}{Hue, Value in the range 0,1}

\item{b}{Brightness}

\item{s}{Saturation}
}
\description{
Function for color based on Hue, Brightness and Sat
}
