% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA_CPOplot.R
\name{INLA_CPOplot}
\alias{INLA_CPOplot}
\title{Function to summarise criteria from an INLA model}
\usage{
INLA_CPOplot(m)
}
\arguments{
\item{m}{INLA model to summarise}
}
\description{
Function to summarise criteria from an INLA model
}
