% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA_CritSum.R
\name{INLA_CritSum}
\alias{INLA_CritSum}
\title{Function to summarise criteria from an INLA model}
\usage{
INLA_CritSum(x)
}
\arguments{
\item{x}{INLA model to summarise}
}
\description{
Function to summarise criteria from an INLA model
}
