% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAUS_CA}
\alias{LAUS_CA}
\title{Local Area Unemployment Statistics.}
\format{
A data frame with seven variables: \code{County}, \code{Date},
\code{Year}, \code{LaborForce}, \code{Employment}, \code{Unemployment} and \code{UnemploymentRate}
}
\usage{
LAUS_CA
}
\description{
Data regarding unemployment rates from 1990 to 2022
}
\details{
Source: https://data.ca.gov/dataset/local-area-unemployment-statistics-laus/resource/b4bc4656-7866-420f-8d87-4eda4c9996ed
}
\keyword{datasets}
