% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalMI.MC.R
\name{LocalMI.MC}
\alias{LocalMI.MC}
\title{Function to obtain pseudo p-values for Morans I}
\usage{
LocalMI.MC(Data, N_simulations, wl)
}
\arguments{
\item{Data}{A vector of continous values that represent the value at each location}

\item{N_simulations}{Number of simulations}

\item{wl}{The adjacency matrix}
}
\description{
Function to obtain pseudo p-values for Morans I
}
