% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.CO.R
\name{MI.CO}
\alias{MI.CO}
\title{Function to obtain pseudo p-values for Morans I}
\usage{
MI.CO(x, var, nsim = 1000)
}
\arguments{
\item{x}{A spatial object of type sf}

\item{var}{The name of a variable from the spatial object which will be used to the Moran's I}

\item{nsim}{Number of simulations}
}
\description{
Function to obtain pseudo p-values for Morans I
}
