% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NA_ufos}
\alias{NA_ufos}
\title{UFO Sightings in North america}
\format{
A data frame with seven variables: \code{County}, \code{Date},
\code{country}, \code{city}, \code{state}, \code{date} and \code{month}
}
\usage{
NA_ufos
}
\description{
Data regarding UFO reports in Canada, Mexico and US
}
\details{
Source: https://nuforc.org
}
\keyword{datasets}
