% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pablette.R
\name{pablette}
\alias{pablette}
\title{Function to get a color palette}
\usage{
pablette(n, pal, type = "cat")
}
\arguments{
\item{n}{number of colors to get}

\item{pal}{palette used, this can be either a vector of colors or a predefined palette}

\item{type}{type of palette, only 'cat' is implemented}
}
\description{
Function to get a color palette
}
