% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cat.R
\name{plot_cat}
\alias{plot_cat}
\title{Function to get the shappiro test p value}
\usage{
plot_cat(x, data, col = "red4", main = NULL, stratified = F, st)
}
\arguments{
\item{x}{name of the variable}

\item{data}{dataset}

\item{col}{color for the fill}

\item{main}{Main title for the figure}
}
\description{
Function to get the shappiro test p value
}
