% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChain.R
\name{plotChain}
\alias{plotChain}
\title{Function to plot a MCMC chain from simulated data}
\usage{
plotChain(df, var, col = "black", lwd = 0.7)
}
\arguments{
\item{df}{the data frame with the simulation outcomes}

\item{var}{A string with the name of the variable to plot}

\item{col}{color of the line to plot}

\item{lwd}{line width for the plot}
}
\description{
Function to plot a MCMC chain from simulated data
}
