% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTheme.R
\name{pTheme}
\alias{pTheme}
\title{Custom theme function}
\usage{
pTheme(dark = T, axis = T, grid = T, legend = "bottom")
}
\arguments{
\item{dark}{boolean, for dark color palettes}

\item{axis}{boolean, whether or not to include axes}

\item{grid}{boolean, whether or not to include the grid}

\item{legend}{legend position, option include: 'top', 'bottom', 'left', 'right', and 'none'}
}
\description{
Custom theme function
}
