% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheets.R
\name{read_sheets}
\alias{read_sheets}
\title{Function to read all the sheets from an excel file into a list}
\usage{
read_sheets(f)
}
\arguments{
\item{f}{the path of the file}
}
\description{
Function to read all the sheets from an excel file into a list
}
