% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadDirs.R
\name{ReadDirs}
\alias{ReadDirs}
\title{Function to read all the files in the directories}
\usage{
ReadDirs(d)
}
\arguments{
\item{d}{Directory path for the simulations}
}
\description{
Function to read all the files in the directories
}
