% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCART.R
\name{RFCART}
\alias{RFCART}
\title{Function to get RF CART plots}
\usage{
RFCART(data, f, main = "", seed = 1)
}
\arguments{
\item{data}{dataset used}

\item{f}{formula for the RF}

\item{main}{Title used for the plots}

\item{seed}{seed used for replication of results}
}
\description{
Function to get RF CART plots
}
