% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_test.R
\name{similarity_test}
\alias{similarity_test}
\title{This function will take a vector of strings and provide a data frame of the similarity between pair of the elements inside the given vector}
\usage{
similarity_test(x, th = 0.5)
}
\arguments{
\item{x}{vector of strings to make the pairwise comparisons}

\item{th}{threshold used for the analysis. values above the threshold will be provided in the output data frame}
}
\description{
This function will take a vector of strings and provide a data frame of the similarity between pair of the elements inside the given vector
}
