% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSum.R
\name{textSum}
\alias{textSum}
\title{Function to get a summary in a string of text}
\usage{
textSum(x, removeCommon = T)
}
\arguments{
\item{x}{string of text to count}

\item{remmoveCommon}{Remove the common words (i.e. 'a', 'are', 'the', ...)}
}
\description{
Function to get a summary in a string of text
}
