% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Custom theme function}
\usage{
theme_map(
  bg = "#E0F0FF",
  grid = "#00000010",
  axis = T,
  legend.position = "right"
)
}
\arguments{
\item{bg}{color of the panel backgound}

\item{grid}{color of the grid}

\item{axis}{boolean, whether or not to include the axis}

\item{legend.position}{legend position, option include: 'top', 'bottom', 'left', 'right', and 'none'}
}
\description{
Custom theme function
}
