% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_print.R
\name{theme_print}
\alias{theme_print}
\title{Function for custom theme}
\usage{
theme_print(
  fill = NA,
  border = "grey50",
  textColor = "grey90",
  grid = "grey40"
)
}
\arguments{
\item{fill}{color for the fill of the panel}

\item{border}{color of the border}
}
\description{
Function for custom theme
}
