% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniMozaic.R
\name{UniMozaic}
\alias{UniMozaic}
\title{Function to get a univariate mozaic plot}
\usage{
UniMozaic(data, x, palette = "Set1")
}
\arguments{
\item{data}{A data frame with the variables}

\item{x}{a string with the name of the variable}

\item{palette}{color palette for the plot}
}
\description{
Function to get a univariate mozaic plot
}
