% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniV.Tbl.R
\name{UniV.Tbl}
\alias{UniV.Tbl}
\title{Function to get a univariate table}
\usage{
UniV.Tbl(Data, x, type = "categorical")
}
\arguments{
\item{Data}{A data frame with the variables}

\item{x}{a string with the name of the variable}

\item{type}{"categorical" or "continous" depending on the type of variable}
}
\description{
Function to get a univariate table
}
