% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxTbl.R
\name{WilcoxTbl}
\alias{WilcoxTbl}
\title{Function to obtain a wilcoxon rank table}
\usage{
WilcoxTbl(Data, x, y)
}
\arguments{
\item{Data}{The name of the data frame}

\item{x}{Name of variable}

\item{y}{Name of respose variable}
}
\description{
Function to obtain a wilcoxon rank table
}
