% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCount.R
\name{wordCount}
\alias{wordCount}
\title{Function to Count the words in a string of text}
\usage{
wordCount(x)
}
\arguments{
\item{x}{string of text to count}
}
\description{
Function to Count the words in a string of text
}
