% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFreq.R
\name{wordFreq}
\alias{wordFreq}
\title{Function to get the word frequency in a string of text}
\usage{
wordFreq(x, removeCommon = T)
}
\arguments{
\item{x}{string of text to count}

\item{remmoveCommon}{Remove the common words (i.e. 'a', 'are', 'the', ...)}
}
\description{
Function to get the word frequency in a string of text
}
